requirements.txt
absl-py==2.1.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
brisque==0.0.16
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
deepface==0.0.75
Django==4.1.3
djangorestframework==3.14.0
executing==2.0.1
filelock==3.15.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
gdown==5.2.0
google-auth==2.33.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.65.4
h5py==3.11.0
humanfriendly==10.0
idna==3.7
imageio==2.35.0
importlib_metadata==8.2.0
importlib_resources==6.4.0
imutils==0.5.4
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
libsvm-official==3.32.0
llvmlite==0.41.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.6.2
matplotlib-inline==0.1.7
mpmath==1.3.0
mtcnn==0.1.1
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
onnxruntime==1.18.1
opencv-python==4.6.0.66
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==3.19.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
Pygments==2.18.0
PyMatting==1.1.12
pyparsing==3.1.2
pyproject-toml==0.0.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyWavelets==1.4.1
referencing==0.35.1
rembg==2.0.58
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rpds-py==0.20.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
sympy==1.13.2
# TA-Lib==0.4.32
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
tifffile==2023.7.10
toml==0.10.2
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.20.0
django-cors-headers
